requirements.txt
pillow==10.3.0
tensorflow==2.16.1
numpy==1.26.4